refactor: 프로젝트 목업데이터 분리에 따른 로직 수정
